+++ src/node/server/index.ts
@@ -64,12 +64,12 @@ export function createServer(config: ServerConfig = {}): Server {
     moduleRewritePlugin,
     moduleResolvePlugin,
     vuePlugin,
+    cssPlugin,
+    ...(transforms.length ? [createServerTransformPlugin(transforms)] : []),
     esbuildPlugin,
     jsonPlugin,
-    cssPlugin,
     htmlPlugin,
     assetPathPlugin,
-    ...(transforms.length ? [createServerTransformPlugin(transforms)] : []),
     serveStaticPlugin
   ]
   resolvedPlugins.forEach((m) => m(context))