+++ src/node/server/serverPluginModuleRewrite.ts
@@ -126,7 +126,7 @@ export function rewriteImports(
       importeeMap.set(importer, currentImportees)
 
       for (let i = 0; i < imports.length; i++) {
-        const { s: start, e: end, d: dynamicIndex } = imports[i]
+        const { s: start, e: end, ss, se, d: dynamicIndex } = imports[i]
         let id = source.substring(start, end)
         let hasLiteralDynamicId = false
         if (dynamicIndex >= 0) {
@@ -145,7 +145,10 @@ export function rewriteImports(
           let resolved
           if (id === hmrClientId) {
             resolved = hmrClientPublicPath
-            if (!/.vue$|.vue\?type=/.test(importer)) {
+            if (
+              /\bhot\b/.test(source.substring(ss, se)) &&
+              !/.vue$|.vue\?type=/.test(importer)
+            ) {
               // the user explicit imports the HMR API in a js file
               // making the module hot.
               rewriteFileWithHMR(root, source, importer, resolver, s)
@@ -249,7 +252,8 @@ export const resolveImport = (
     }
 
     // mark non-src imports
-    if (!jsSrcRE.test(pathname)) {
+    const ext = path.extname(pathname)
+    if (ext && !jsSrcRE.test(pathname)) {
       query += `${query ? `&` : `?`}import`
     }