+++ src/node/server/serverPluginVue.ts
@@ -31,6 +31,7 @@ import { transform } from '../esbuildService'
 import { InternalResolver } from '../resolver'
 import qs from 'querystring'
 import { seenUrls } from './serverPluginServeStatic'
+import { rewriteCssUrls } from '../utils/cssUtils'
 
 const debug = require('debug')('vite:sfc')
 const getEtag = require('etag')
@@ -436,6 +437,9 @@ async function compileSFCStyle(
       }
     })
   }
+
+  // rewrite relative urls
+  result.code = await rewriteCssUrls(result.code, publicPath)
 
   cached = cached || { styles: [] }
   cached.styles[index] = result